+++ test/clojure-mode-indentation-test.el
@@ -421,7 +421,9 @@ values of customisable variables."
 (def-full-indent-test let-when-while-forms
   "(let-alist [x 1]\n  ())"
   "(while-alist [x 1]\n  ())"
-  "(when-alist [x 1]\n  ())")
+  "(when-alist [x 1]\n  ())"
+  "(if-alist [x 1]\n  ())"
+  "(indents-like-fn-when-let-while-if-are-not-the-start [x 1]\n                                                     ())")
 
 (defun indent-cond (indent-point state)
   (goto-char (elt state 1))